+++ pkg/unittest/formatter/sonar_report_xml_test.go
@@ -2,6 +2,7 @@ package formatter_test
 
 import (
 	"encoding/xml"
+	"fmt"
 	"path/filepath"
 	"testing"
 	"time"
@@ -11,18 +12,31 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func createSonarTestCase(name string, duration string, isError bool, isFailed bool) SonarTestCase {
+func createSonarTestCase(name, duration, message, reason string, isSkipped, isError, isFailed bool) SonarTestCase {
 	testCase := SonarTestCase{
 		Name:     name,
 		Duration: duration,
 	}
 
+	if isSkipped {
+		testCase.Skipped = &SonarSkipped{
+			Message: message,
+			Reason:  reason,
+		}
+	}
+
 	if isError {
-		testCase.Error = &SonarError{}
+		testCase.Error = &SonarError{
+			Message:    message,
+			Stacktrace: reason,
+		}
 	}
 
 	if isFailed {
-		testCase.Failure = &SonarFailure{}
+		testCase.Failure = &SonarFailure{
+			Message:    message,
+			Stacktrace: reason,
+		}
 	}
 
 	return testCase
@@ -44,12 +58,35 @@ func validateSonarTestCases(assert *assert.Assertions, expected, actual []SonarT
 		assert.Equal(expected[i].Name, actual[i].Name)
 		assert.Equal(expected[i].Duration, actual[i].Duration)
 
+		if expected[i].Skipped != nil {
+			assert.NotNil(actual[i].Skipped)
+			assert.Equal(expected[i].Skipped.Message, actual[i].Skipped.Message)
+			assert.Equal(expected[i].Skipped.Reason, actual[i].Skipped.Reason)
+		} else {
+			assert.Nil(actual[i].Skipped)
+		}
+
+		if expected[i].Error != nil {
+			assert.NotNil(actual[i].Error)
+			assert.Equal(expected[i].Error.Message, actual[i].Error.Message)
+			assert.Equal(expected[i].Error.Stacktrace, actual[i].Error.Stacktrace)
+		} else {
+			assert.Nil(actual[i].Error)
+		}
+
+		if expected[i].Failure != nil {
+			assert.NotNil(actual[i].Failure)
+			assert.Equal(expected[i].Failure.Message, actual[i].Failure.Message)
+			assert.Equal(expected[i].Failure.Stacktrace, actual[i].Failure.Stacktrace)
+		} else {
+			assert.Nil(actual[i].Failure)
+		}
 	}
 }
 
 func TestWriteTestOutputAsSonarNoTests(t *testing.T) {
 	assert := assert.New(t)
-	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_Test_Output.xml")
+	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_NoTests_Output.xml")
 
 	expected := SonarTestExecutions{
 		Version: 1,
@@ -70,7 +107,7 @@ func TestWriteTestOutputAsSonarNoTests(t *testing.T) {
 
 func TestWriteTestOutputAsSonarMinimalSuccess(t *testing.T) {
 	assert := assert.New(t)
-	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_Test_Output.xml")
+	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_MinimalSuccess_Output.xml")
 	testSuiteDisplayName := "TestingSuiteSuccess"
 	testCaseDisplayName := "TestCaseSuccessSuccess"
 
@@ -82,6 +119,9 @@ func TestWriteTestOutputAsSonarMinimalSuccess(t *testing.T) {
 					createSonarTestCase(
 						testCaseDisplayName,
 						"12000",
+						"",
+						"",
+						false,
 						false,
 						false,
 					),
@@ -97,7 +137,7 @@ func TestWriteTestOutputAsSonarMinimalSuccess(t *testing.T) {
 			FilePath:    outputFile,
 			Passed:      true,
 			TestsResult: []*results.TestJobResult{
-				createTestJobResult(testCaseDisplayName, "", true, nil),
+				createTestJobResult(testCaseDisplayName, "", true, false, nil),
 			},
 		},
 	}
@@ -116,12 +156,13 @@ func TestWriteTestOutputAsSonarMinimalSuccess(t *testing.T) {
 
 func TestWriteTestOutputAsSonarWithErrorsAndFailures(t *testing.T) {
 	assert := assert.New(t)
-	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_Test_Output.xml")
+	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_ErrorsAndFailures_Output.xml")
 	testSuiteDisplayName := "TestingSuiteFailuresAndErrors"
 	testCaseDisplayNameError := "TestCaseError"
 	testCaseDisplayNameFailure := "TestCaseFailure"
 	assertionFailure := "AssertionFailure"
 	assertionType := "equal"
+	failureMessage := fmt.Sprintf("\t\t - asserts[0] `%s` fail \n\t\t\t %s \n", assertionType, assertionFailure)
 
 	expected := SonarTestExecutions{
 		Files: []SonarFile{
@@ -131,12 +172,18 @@ func TestWriteTestOutputAsSonarWithErrorsAndFailures(t *testing.T) {
 					createSonarTestCase(
 						testCaseDisplayNameError,
 						"123",
+						"Error",
+						"TheError",
+						false,
 						true,
 						false,
 					),
 					createSonarTestCase(
 						testCaseDisplayNameFailure,
 						"456",
+						"Failed",
+						failureMessage,
+						false,
 						false,
 						true,
 					),
@@ -147,7 +194,7 @@ func TestWriteTestOutputAsSonarWithErrorsAndFailures(t *testing.T) {
 	}
 
 	assertionResults := []*results.AssertionResult{
-		createAssertionResult(0, false, false, assertionType, assertionFailure, ""),
+		createAssertionResult(0, false, false, false, assertionType, assertionFailure, "", ""),
 	}
 
 	given := []*results.TestSuiteResult{
@@ -156,13 +203,70 @@ func TestWriteTestOutputAsSonarWithErrorsAndFailures(t *testing.T) {
 			FilePath:    outputFile,
 			Passed:      true,
 			TestsResult: []*results.TestJobResult{
-				createTestJobResult(testCaseDisplayNameError, "TheError", false, nil),
-				createTestJobResult(testCaseDisplayNameFailure, "", false, assertionResults),
+				createTestJobResult(testCaseDisplayNameError, "TheError", false, false, nil),
+				createTestJobResult(testCaseDisplayNameFailure, "", false, false, assertionResults),
 			},
 		},
 	}
 	given[0].TestsResult[0].Duration, _ = time.ParseDuration("123ms")
 	given[0].TestsResult[1].Duration, _ = time.ParseDuration("456ms")
+
+	sut := NewSonarReportXML()
+	byteValue := loadFormatterTestcase(assert, outputFile, given, sut)
+
+	var actual SonarTestExecutions
+	err := xml.Unmarshal(byteValue, &actual)
+	assert.Nil(err)
+
+	assert.Equal(expected.Version, actual.Version)
+	validateSonarFiles(assert, expected.Files, actual.Files)
+}
+
+func TestWriteTestOutputAsSonarWithSkipped(t *testing.T) {
+	assert := assert.New(t)
+	outputFile := filepath.Join(tmpNunitTestDir, "Sonar_Skipped_Output.xml")
+	testSuiteDisplayName := "TestingSuiteSkipped"
+	testCaseDisplayNameSkipped := "TestCaseSkipped"
+	assertionType := "equal"
+	assertIndex := 0
+	skipReason := "Version mismatch"
+	skippedContent := fmt.Sprintf("SKIPPED '%s' \n\t\t - asserts[%v] `%s` fail \n\t\t\t %s \n", testCaseDisplayNameSkipped, assertIndex, assertionType, skipReason)
+
+	expected := SonarTestExecutions{
+		Files: []SonarFile{
+			{
+				Path: outputFile,
+				TestCases: []SonarTestCase{
+					createSonarTestCase(
+						testCaseDisplayNameSkipped,
+						"1",
+						"Skipped",
+						skippedContent,
+						true,
+						false,
+						false,
+					),
+				},
+			},
+		},
+		Version: 1,
+	}
+
+	assertionSkippedResults := []*results.AssertionResult{
+		createAssertionResult(assertIndex, false, true, false, assertionType, "", skipReason, ""),
+	}
+
+	given := []*results.TestSuiteResult{
+		{
+			DisplayName: testSuiteDisplayName,
+			FilePath:    outputFile,
+			Passed:      true,
+			TestsResult: []*results.TestJobResult{
+				createTestJobResult(testCaseDisplayNameSkipped, "", false, true, assertionSkippedResults),
+			},
+		},
+	}
+	given[0].TestsResult[0].Duration, _ = time.ParseDuration("1ms")
 
 	sut := NewSonarReportXML()
 	byteValue := loadFormatterTestcase(assert, outputFile, given, sut)