+++ pkg/unittest/test_suite.go
@@ -212,6 +212,7 @@ type TestSuite struct {
 	Set              map[string]any
 	Templates        []string
 	ExcludeTemplates []string `yaml:"excludeTemplates"`
+	IncludeCrds      bool     `yaml:"includeCrds"`
 	Release          struct {
 		Name      string
 		Namespace string
@@ -238,6 +239,8 @@ type TestSuite struct {
 	chartRoute string
 	// if true, indicates that this was created from a helm rendered file
 	fromRender bool
+	// if true, skip values.schema.json validation when rendering
+	skipSchemaValidation bool
 	// An identifier to append to snapshot files
 	SnapshotId string `yaml:"snapshotId"`
 	Skip       struct {
@@ -409,6 +412,8 @@ func (s *TestSuite) runV3TestJobs(
 				WithFailFast(failFast),
 				WithPostRendererConfig(s.PostRendererConfig),
 				WithDocumentSelector(testJob.DocumentSelector),
+				WithIncludeCrds(s.IncludeCrds),
+				WithSkipSchemaValidation(s.skipSchemaValidation),
 			))
 			jobResult = testJob.RunV3(&job)
 			jobResults[idx] = jobResult